requirements.txt
flask
google-api-python-client
google-auth-oauthlib
python-dateutil
google-generativeai
scikit-learn
joblib
python-dotenv